+++ Vagrantfile
@@ -6,18 +6,28 @@ Vagrant.configure("2") do |config|
   ansible_extra_vars['local_repository'] = '/vagrant'
   ansible_extra_vars['log_level'] = ENV['LOG_LEVEL'] || 'DEBUG'
 
-  if ENV.has_key? 'SLACK_APP_TOKEN'
+  %w(YasOpenStack YasExampleHandlers).each do |repo|
+    local_path = '../' + repo
+    remote_path = '/srv/' + repo
+    if Dir.exists?(local_path) then
+      config.vm.synced_folder(local_path, remote_path)
+    end
+  end
+
+  if ENV.has_key?('SLACK_APP_TOKEN')
     ansible_extra_vars['slack_app_token'] = ENV['SLACK_APP_TOKEN']
   end
 
-  if ENV.has_key? 'SLACK_APP_NAME'
+  if ENV.has_key?('SLACK_APP_NAME')
     ansible_extra_vars['slack_app_name'] = ENV['SLACK_APP_NAME']
   end
 
-  config.vm.provision :ansible,
+  config.vm.provision(
+    :ansible,
     playbook: "playbooks/main.yml",
     sudo: true,
     verbose: "vv",
     raw_arguments: ["--diff"],
     extra_vars: ansible_extra_vars
+  )
 end
