+++ setup.cfg
@@ -24,4 +24,4 @@ data_files =
 
 [entry_points]
 console_scripts =
-    yas = yas.application:run
+    yas = yas.core.application:run
