+++ yas/core/client.py
@@ -4,12 +4,10 @@ from threading import Thread
 
 from slackclient import SlackClient
 
-from yas.handler_manager import HandlerManager
-from yas.logging import logger
-from yas.yaml_file_config import YamlConfiguration
+from yas.core.handler_manager import HandlerManager
+from yas.core.logging import Logger
+from yas.core.yaml_file_config import YamlConfiguration
 
-
-CONFIG = YamlConfiguration()
 
 def rough_hash(string):
     return hashlib.md5(str(string).encode('utf-8')).hexdigest()
@@ -17,42 +15,38 @@ def rough_hash(string):
 class Client(SlackClient):
 
     def __init__(self, ignored_types=None):
-        super().__init__(CONFIG.slack_app_token)
-
-        self.ignored_types = ignored_types or CONFIG.ignored_types
-
-        self.handler_manager = HandlerManager(
-            CONFIG.handler_list,
-            CONFIG.bot_name,
-            self.api_call,
-            debug=CONFIG.debug)
+        self.config = YamlConfiguration()
+        self.log = Logger(self.config.log_level).log
+        super().__init__(self.config.slack_app_token)
+        self.handler_manager = HandlerManager(self.config, self.api_call, self.log)
 
     def process_changes(self, data):
         super().process_changes(data)
-        logger.log.debug(f"Raw data: {data}")
+        self.log.debug(f"Raw data: {data}")
         channel = data.get('channel')
         def reply(message, channel=channel, thread=None, reply_broadcast=None):
             if channel:
                 self.rtm_send_message(channel, message, thread, reply_broadcast)
         data['yas_hash'] = rough_hash(data)
-        logger.log.info(f"Processing: {data}")
+        self.log.info(f"Processing: {data}")
         try:
             thread = Thread(target=self.handler_manager.handle, args=(data, reply))
             thread.start()
         # pylint: disable=broad-except
         except Exception as exception:
-            reply(CONFIG.handler_exception_message.format(exception=exception))
+            reply(self.config.handler_exception_message.format(exception=exception))
             # pylint: disable=raising-bad-type
             raise exception
 
     def listen(self):
         if self.rtm_connect():
-            logger.log.info(f"Slack bot connected as {CONFIG.bot_name} and running!")
+            self.log.info(f"Slack bot connected as {self.config.bot_name} and running!")
             while True:
                 self.rtm_read()
                 time.sleep(0.01)
         else:
-            logger.log.fatal(
-                "Connection failed. If this occured after the bot had been running"
-                "a while, it was probably just a connection blip; otherwise, confirm"
-                "the yas.yml has a valid token and name.")
+            self.log.fatal(
+                "Connection failed. If this occured after the bot had been running "
+                "a while, it was probably just a connection blip; otherwise, confirm "
+                "the yas.yml has a valid token and name."
+            )
