+++ yas/handlers/identify_handler.py
@@ -1,17 +1,14 @@
 import socket
 
 from yas import RegexHandler
-from yas.yaml_file_config import YamlConfiguration
 
-
-CONFIG = YamlConfiguration()
 
 class IdentifyHandler(RegexHandler):
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(r'id', *args, **kwargs)
+    def __init__(self, bot):
+        super().__init__(r'id', bot)
 
-    def handle(self, data, reply):
+    def handle(self, _, reply):
         soc = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
         soc.connect(("8.8.8.8", 80))
         my_ip = soc.getsockname()[0]
