+++ yas/handlers/ignored_types_handler.py
@@ -1,0 +1,10 @@
+from yas import YasHandler
+
+
+class IgnoredTypesHandler(YasHandler):
+    def test(self, data):
+        if data.get('type') in self.bot.config.ignored_types:
+            return True
+
+    def handle(self, data, _):
+        pass
